Add confirmation dialog for user logout
