+++ app/src/main/java/org/stepic/droid/view/dialogs/AreYouSureDialog.java
@@ -1,8 +1,8 @@
 package org.stepic.droid.view.dialogs;
 
 import android.app.Dialog;
-import android.app.DownloadManager;
 import android.content.DialogInterface;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
 import android.support.v7.app.AlertDialog;
@@ -12,24 +12,20 @@ import com.yandex.metrica.YandexMetrica;
 import org.jetbrains.annotations.NotNull;
 import org.stepic.droid.R;
 import org.stepic.droid.base.MainApplication;
-import org.stepic.droid.model.DownloadEntity;
-import org.stepic.droid.preferences.UserPreferences;
+import org.stepic.droid.core.IShell;
+import org.stepic.droid.preferences.SharedPreferenceHelper;
 import org.stepic.droid.store.operations.DatabaseManager;
 import org.stepic.droid.util.AppConstants;
-import org.stepic.droid.util.CleanerUtil;
-
-import java.util.List;
 
 import javax.inject.Inject;
 
-public class ClearCacheDialogFragment extends DialogFragment {
+public class AreYouSureDialog extends DialogFragment {
 
     @Inject
-    DatabaseManager mDatabaseManager;
+    IShell mShell;
+
     @Inject
-    UserPreferences userPreferences;
-    @Inject
-    DownloadManager mSystemDownloadManager;
+    DatabaseManager mDbManager;
 
     @NotNull
     @Override
@@ -38,20 +34,24 @@ public class ClearCacheDialogFragment extends DialogFragment {
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity(), R.style.MyAlertDialogStyle);
         builder.setTitle(R.string.title_clear_cache_dialog)
-                .setMessage(R.string.are_you_sure)
+                .setMessage(R.string.clear_cache_dialog_message)
                 .setPositiveButton(R.string.yes, new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        YandexMetrica.reportEvent(AppConstants.METRICA_CLICK_YES_CLEAR_CACHE);
-                        //// FIXME: 22.10.15 do it in background
-                        List<DownloadEntity> downloadEntities = mDatabaseManager.getAllDownloadEntities();
-                        for (DownloadEntity de : downloadEntities) {
-                            mSystemDownloadManager.remove(de.getDownloadId());
-                        }
+                        YandexMetrica.reportEvent(AppConstants.METRICA_CLICK_YES_LOGOUT);
 
-                        CleanerUtil.CleanDirectory(userPreferences.getDownloadFolder());
 
-                        mDatabaseManager.dropDatabase();
+                        SharedPreferenceHelper helper = mShell.getSharedPreferenceHelper();
+                        helper.deleteAuthInfo();
+                        AsyncTask<Void, Void, Void> task = new AsyncTask<Void, Void, Void>() {
+                            @Override
+                            protected Void doInBackground(Void... params) {
+                                mDbManager.clearCacheCourses(DatabaseManager.Table.enrolled);
+                                return null;
+                            }
+                        };
+                        task.execute();
+                        mShell.getScreenProvider().showLaunchScreen(MainApplication.getAppContext(), false);
                     }
                 })
                 .setNegativeButton(R.string.no, null);
